requirements.txt
Django
pillow
django-tinymce4-widget
psycopg2
